Move SettingsTab out of ui.py
